ToDo Recycler on Room фильтр обновляется Add Delete итд  работает Clear create in menuBar количество записей в заголовке Repository ToDo03 Card + Menu DataConverters Canary 8 Date Int + Char Index MemoryBase Old+Pager_Cheese Pager Ready ToDo SQL PAEMI - Int Id Int Calendar Refactor and and ToDo_SQL Enum 8 млн

+++ ToDo/app/src/main/java/com/app4web/asdzendo/todo/launcher/ToDoActitityViewModel.kt
@@ -15,5 +15,5 @@ class ToDoActitityViewModel internal constructor(
 
     fun count() = factRepository.count()
 
-    fun addFactDatabase(COUNTSFact: Long) = factRepository.addFactDatabase(COUNTSFact)
+    fun addFactDatabase(COUNTSFact: Int) = factRepository.addFactDatabase(COUNTSFact)
 }
